+++ src/lib/components/index.ts
@@ -1,0 +1,7 @@
+export * from "./toolbar";
+export * from "./process";
+export * from "./stats";
+export * from "./modals";
+export { default as AppInfo } from "./AppInfo.svelte";
+export { default as TitleBar } from "./TitleBar.svelte";
+export { default as ThemeSwitcher } from "./ThemeSwitcher.svelte";
