+++ src/lib/components/stats/NetworkPanel.svelte
@@ -1,23 +1,10 @@
 <script lang="ts">
   import { faNetworkWired } from "@fortawesome/free-solid-svg-icons";
-  import PanelHeader from "./PanelHeader.svelte";
-  import StatItem from "./StatItem.svelte";
+  import { PanelHeader, StatItem } from "$lib/components";
+  import { formatBytes } from "$lib/utils";
 
   export let networkRxBytes: number;
   export let networkTxBytes: number;
-
-  function formatBytes(bytes: number): string {
-    const units = ["B", "KB", "MB", "GB", "TB"];
-    let value = bytes;
-    let unitIndex = 0;
-
-    while (value >= 1024 && unitIndex < units.length - 1) {
-      value /= 1024;
-      unitIndex++;
-    }
-
-    return `${value.toFixed(1)} ${units[unitIndex]}/s`;
-  }
 </script>
 
 <div class="stat-panel">
