--- src/lib/constants/toolbar.ts
@@ -1,9 +1,0 @@
-export const ITEMS_PER_PAGE_OPTIONS = [15, 25, 50, 100, 250, 500];
-
-export const REFRESH_RATE_OPTIONS = [
-  { value: 1000, label: "1s" },
-  { value: 2000, label: "2s" },
-  { value: 5000, label: "5s" },
-  { value: 10000, label: "10s" },
-  { value: 30000, label: "30s" },
-];
