+++ src/lib/definitions/themes.ts
@@ -1,30 +1,4 @@
-export interface Theme {
-  name: string;
-  label: string;
-  colors: {
-    base: string;
-    mantle: string;
-    crust: string;
-    text: string;
-    subtext0: string;
-    subtext1: string;
-    surface0: string;
-    surface1: string;
-    surface2: string;
-    overlay0: string;
-    overlay1: string;
-    blue: string;
-    lavender: string;
-    sapphire: string;
-    sky: string;
-    red: string;
-    maroon: string;
-    peach: string;
-    yellow: string;
-    green: string;
-    teal: string;
-  };
-}
+import type { Theme } from "$lib/types";
 
 export const themes: Record<string, Theme> = {
   catppuccin: {
